screen-orientation/onchange-event-subframe.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<iframe id="testIframe" sandbox="allow-scripts allow-same-origin"
  style="display:none" src="resources/iframe-listen-orientation-change.html">
</iframe>

<script>
promise_test(async t => {
  const orientations = [
    'portrait-primary',
    'portrait-secondary',
    'landscape-primary',
    'landscape-secondary'
  ];
  const messageWatcher = new EventWatcher(t, window, "message");

  let currentIndex = orientations.indexOf(screen.orientation.type);
  let nextIndex;
  let message;
  for (let i = 0; i < orientations.length; i++) {
    nextIndex = (currentIndex + 1) % orientations.length;
    screen.orientation.lock(orientations[nextIndex]);
    currentIndex = nextIndex;
    message = await messageWatcher.wait_for("message");
    assert_equals(message.data, orientations[currentIndex], "subframe receives orientation change event");
  }
  screen.orientation.unlock();
}, "Test subframes receive orientation change events");
</script>
